+++ tests/Parsers/EutilsAssemblyXMLParserTest.php
@@ -5,9 +5,8 @@ namespace Tests;
 use StatonLab\TripalTestSuite\DBTransaction;
 use StatonLab\TripalTestSuite\TripalTestCase;
 
-class EUtilsXmlParserTest extends TripalTestCase {
+class EutilsAssemblyXMLParserTest extends TripalTestCase {
 
-  // Uncomment to auto start and rollback db transactions per test method.
   use DBTransaction;
 
   /**
@@ -15,31 +14,6 @@ class EUtilsXmlParserTest extends TripalTestCase {
    */
   protected $assembly_xmls = NULL;
 
-
-  /**
-   * @test
-   * @group biosamples
-   */
-  public function testBioSampleParser() {
-    $path = DRUPAL_ROOT . '/' . drupal_get_path('module', 'tripal_eutils');
-    foreach (glob("$path/examples/biosamples/*.xml") as $file) {
-      $parser = new \EUtilsBioSampleParser();
-      $biosample = $parser->parse(simplexml_load_file($file));
-
-      $this->assertArrayHasKey('name', $biosample);
-      $this->assertArrayHasKey('accessions', $biosample);
-      $this->assertArrayHasKey('attributes', $biosample);
-      $this->assertArrayHasKey('description', $biosample);
-      $this->assertTrue(is_array($biosample['accessions']));
-      $this->assertTrue(is_array($biosample['attributes']));
-
-      if ($biosample['name'] == 'SAMN02953603'){
-        // This sample has a linked project.  Was it parsed?
-        $this->assertArrayHasKey('projects', $biosample);
-        $this->assertNotEmpty($biosample['projects']);
-      }
-    }
-  }
 
   /**
    * @group assembly
